+++ client/src/services/runAllVerifiers.mts
