+++ client/src/services/verifyResults.mts
